requirements.txt
alembic==0.9.9
amqp==2.2.2
astroid==1.6.2
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Celery-Helper==1.1.0
Flask-Email==1.4.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httplib2==0.11.3
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
oauth2==1.9.0.post1
pycparser==2.18
pylint==1.8.3
PyMySQL==0.8.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.5
vine==1.1.4
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
